+++ YWXiMaLaYa/YWXiMaLaYa/Class/PublicClass/Player/YWRequestTask.m
@@ -29,7 +29,7 @@
 - (void)start {
     NSMutableURLRequest * request = [NSMutableURLRequest requestWithURL:[self.requestURL originalSchemeURL] cachePolicy:NSURLRequestReloadIgnoringCacheData timeoutInterval:RequestTimeout];
     if (self.requestOffset > 0) {
-        [request addValue:[NSString stringWithFormat:@"bytes=%ld-%ld", self.requestOffset, self.fileLength - 1] forHTTPHeaderField:@"Range"];
+        [request addValue:[NSString stringWithFormat:@"bytes=%ld-%lu", (unsigned long)self.requestOffset, self.fileLength - 1] forHTTPHeaderField:@"Range"];
     }
     self.session = [NSURLSession sessionWithConfiguration:[NSURLSessionConfiguration defaultSessionConfiguration] delegate:self delegateQueue:[NSOperationQueue mainQueue]];
     self.task = [self.session dataTaskWithRequest:request];
